refactor: break up large testfiles
