+++ internal/credit/balancesnapshot_connector.go
@@ -8,13 +8,13 @@ import (
 	"github.com/openmeterio/openmeter/pkg/framework/entutils"
 )
 
-type BalanceSnapshotConnector interface {
+type BalanceSnapshotRepo interface {
 	InvalidateAfter(ctx context.Context, owner NamespacedGrantOwner, at time.Time) error
 	GetLatestValidAt(ctx context.Context, owner NamespacedGrantOwner, at time.Time) (GrantBalanceSnapshot, error)
 	Save(ctx context.Context, owner NamespacedGrantOwner, balances []GrantBalanceSnapshot) error
 
 	entutils.TxCreator
-	entutils.TxUser[BalanceSnapshotConnector]
+	entutils.TxUser[BalanceSnapshotRepo]
 }
 
 // No balance has been saved since start of measurement for the owner
