+++ test/entitlement/regression/framework_test.go
@@ -27,7 +27,7 @@ import (
 type Dependencies struct {
 	GrantRepo                credit.GrantRepo
 	GrantDB                  *grantdb.Client
-	BalanceSnapshotConnector credit.BalanceSnapshotConnector
+	BalanceSnapshotConnector credit.BalanceSnapshotRepo
 	GrantConnector           credit.GrantConnector
 
 	EntitlementRepo entitlement.EntitlementRepo
